mindease-flask/requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-CORS==4.0.0
python-dotenv==1.0.0
google-generativeai>=0.5.0
requests>=2.31.0
